Fix signed offset arithmetic and line wrapping in VM handlers
- offsetPtrHandler: cast to int64_t before multiplying int32_t offset
  by uint32_t opcodeExtension to preserve sign (pre-existing bug)
- getElementHandler: wrap long memcpy call for formatting

+++ source/slang/slang-vm-inst-impl.cpp
@@ -592,7 +592,7 @@ static void offsetPtrHandler(IByteCodeRunner* ctx, VMExecInstHeader* inst, void*
     memcpy(&basePtr, inst->getOperand(1).getPtr(), sizeof(void*));
     int32_t offset;
     memcpy(&offset, inst->getOperand(2).getPtr(), sizeof(offset));
-    void* result = basePtr + offset * inst->opcodeExtension;
+    void* result = basePtr + (int64_t)offset * (int64_t)inst->opcodeExtension;
     memcpy(inst->getOperand(0).getPtr(), &result, sizeof(void*));
 }
